Rename configupdater package to fluxupdater
