+++ pkg/fluxupdater/fluxupdater_test.go
@@ -1,4 +1,4 @@
-package configupdater
+package fluxupdater
 
 import (
 	"fmt"
